requirements.txt
django
django-markdownx